Use float instead of double for the PID path
The PID library's author recommends this explicitly (see the library
commit): on AVR a double *is* a float, so the choice never carried
meaning there. On the ESP32 the FPU is single precision only, so every
double operation is emulated in software by libgcc.

Measured on an ESP32-D0WD-V3 at 240 MHz: addition 53.6 vs 4.0 cycles,
multiplication 104.6 vs 4.0, a full PID step 295 vs 26 cycles. The
speed-up is real but not the reason for this change: at a one-second
sample time it saves roughly a microsecond per second.

The reason is flash. The build sits at 90.8 % of a 1703936-byte app
partition, and because OTA needs two equally sized app partitions, that
ceiling cannot simply be raised. 2880 bytes are 1.85 % of what is left,
and about 1.8 kB of it is the libgcc emulation
(__adddf3/__muldf3/__divdf3) that stops being linked in at all.

Changed the PID-adjacent globals in main.cpp (setpoints, tuning
parameters, temperature, pidOutput) and the extern declaration in isr.h.
The latter also makes the heater ISR's read of pidOutput a single 32-bit
access, i.e. atomic; it previously read a 64-bit value that the loop
could be half-way through writing, 100 times per second.

Registry deklarations had to follow: Parameter.cpp casts the global
variable pointer by the declared EditableKind, not by the template type,
so a parameter declared kDouble whose globalVar points at a float would
write 8 bytes into 4 and corrupt whatever follows it. That compiles
cleanly, which makes it worth spelling out. 16 parameters are therefore
now kFloat; every remaining kDouble was checked to point at a real
double or at nullptr.

Untouched on purpose: the brew/backflush timings (still double, and
declared kDouble accordingly) and the ParameterRegistry/Parameter
abstraction itself, which carries its own kDouble type.

Saves 2880 bytes of flash and 200 bytes of RAM.

A note on the atomicity argument, so nobody overstates it: with one
write per second and a 10 ms ISR, a torn read is expected roughly every
two weeks -- real, but it decides a single 10 ms heater window, which is
about 0.01 K on the boiler. Worth having, not a defect being fixed.

NOTE: platformio.ini temporarily points at a local symlink of the PID
library. Before this goes anywhere, it has to point at a published
commit of the library with the float change.

+++ src/ParameterRegistry.cpp
@@ -95,7 +95,7 @@ void ParameterRegistry::initialize(Config& config) {
     addNumericConfigParam<double>(
         "pid.ema_factor",
         "PID EMA Factor",
-        kDouble,
+        kFloat,
         sPIDSection,
         111,
         &emaFactor,
@@ -107,7 +107,7 @@ void ParameterRegistry::initialize(Config& config) {
     addNumericConfigParam<double>(
         "pid.regular.kp",
         "PID Kp",
-        kDouble,
+        kFloat,
         sPIDSection,
         112,
         &aggKp,
@@ -120,7 +120,7 @@ void ParameterRegistry::initialize(Config& config) {
     addNumericConfigParam<double>(
         "pid.regular.tn",
         "PID Tn (=Kp/Ki)",
-        kDouble,
+        kFloat,
         sPIDSection,
         113,
         &aggTn,
@@ -133,7 +133,7 @@ void ParameterRegistry::initialize(Config& config) {
     addNumericConfigParam<double>(
         "pid.regular.tv",
         "PID Tv (=Kd/Kp)",
-        kDouble,
+        kFloat,
         sPIDSection,
         114,
         &aggTv,
@@ -146,7 +146,7 @@ void ParameterRegistry::initialize(Config& config) {
     addNumericConfigParam<double>(
         "pid.regular.i_max",
         "PID Integrator Max",
-        kDouble,
+        kFloat,
         sPIDSection,
         115,
         &aggIMax,
@@ -159,7 +159,7 @@ void ParameterRegistry::initialize(Config& config) {
     addNumericConfigParam<double>(
         "pid.steam.kp",
         "Steam Kp",
-        kDouble,
+        kFloat,
         sPIDSection,
         116,
         &steamKp,
@@ -172,7 +172,7 @@ void ParameterRegistry::initialize(Config& config) {
     addNumericConfigParam<double>(
         "TEMP",
         "Temperature",
-        kDouble,
+        kFloat,
         sTempSection,
         200,
         &temperature,
@@ -185,7 +185,7 @@ void ParameterRegistry::initialize(Config& config) {
     addNumericConfigParam<double>(
         "brew.setpoint",
         "Setpoint (°C)",
-        kDouble,
+        kFloat,
         sTempSection,
         201,
         &brewSetpoint,
@@ -197,7 +197,7 @@ void ParameterRegistry::initialize(Config& config) {
     addNumericConfigParam<double>(
         "brew.temp_offset",
         "Offset (°C)",
-        kDouble,
+        kFloat,
         sTempSection,
         202,
         &brewTempOffset,
@@ -209,7 +209,7 @@ void ParameterRegistry::initialize(Config& config) {
     addNumericConfigParam<double>(
         "steam.setpoint",
         "Steam Setpoint (°C)",
-        kDouble,
+        kFloat,
         sTempSection,
         203,
         &steamSetpoint,
@@ -412,7 +412,7 @@ void ParameterRegistry::initialize(Config& config) {
         addNumericConfigParam<double>(
             "brew.pid_delay",
             "Brew PID Delay (s)",
-            kDouble,
+            kFloat,
             sBrewPidSection,
             711,
             &brewPidDelay,
@@ -424,7 +424,7 @@ void ParameterRegistry::initialize(Config& config) {
         addNumericConfigParam<double>(
             "pid.bd.kp",
             "BD Kp",
-            kDouble,
+            kFloat,
             sBrewPidSection,
             712,
             &aggbKp,
@@ -436,7 +436,7 @@ void ParameterRegistry::initialize(Config& config) {
         addNumericConfigParam<double>(
             "pid.bd.tn",
             "BD Tn (=Kp/Ki)",
-            kDouble,
+            kFloat,
             sBrewPidSection,
             713,
             &aggbTn,
@@ -448,7 +448,7 @@ void ParameterRegistry::initialize(Config& config) {
         addNumericConfigParam<double>(
             "pid.bd.tv",
             "BD Tv (=Kd/Kp)",
-            kDouble,
+            kFloat,
             sBrewPidSection,
             714,
             &aggbTv,
@@ -509,7 +509,7 @@ void ParameterRegistry::initialize(Config& config) {
     addNumericConfigParam<double>(
         "standby.time",
         "Standby Time",
-        kDouble,
+        kFloat,
         sPowerSection,
         802,
         &standbyModeTime,
@@ -598,7 +598,7 @@ void ParameterRegistry::initialize(Config& config) {
     addNumericConfigParam(
         "display.post_brew_timer_duration",
         "Post Brew Timer Duration (s)",
-        kDouble,
+        kFloat,
         sDisplaySection,
         908,
         &postBrewTimerDuration,
